+++ OpenGL Sandbox/DirectionalLight.hpp
@@ -24,7 +24,7 @@ public:
     
     glm::mat4 CalculateLightTransform();
     
-    void UseLight(GLfloat ambientIntensityLocation, GLfloat ambientColourLocation, GLfloat diffuseIntensityLocation, GLfloat directionLocation);
+    void UseLight(GLuint ambientIntensityLocation, GLuint ambientColourLocation, GLuint diffuseIntensityLocation, GLuint directionLocation);
     
 private:
     glm::vec3 direction;
